refactor(types): consolidate src/data/types.ts into root src/types.ts (refs #298)
