+++ src/components/DataTablePanel.tsx
@@ -23,13 +23,12 @@ import { LoadingState, PanelProps, textUtil } from '@grafana/data';
 import { useStyles2, useTheme2 } from '@grafana/ui';
 import { useApplyTransformation } from 'hooks/useApplyTransformation';
 import React, { useEffect, useMemo, useRef, useState } from 'react';
-import { DatatableOptions } from 'types';
+import { DatatableOptions, DTColumnType, DTData } from 'types';
 import { BuildColumnDefs, ConvertDataFrameToDataTableFormat } from 'data/dataHelpers';
 import { ApplyColumnWidthHints } from 'data/columns/columnWidthHints';
 import { buildSearchBarLayout } from 'data/layout/buildSearchBarLayout';
 import { datatableThemedStyles } from './styles';
 import { GetDataTransformerID } from 'data/transformations';
-import { DTColumnType, DTData } from 'data/types';
 import { ApplyColumnAliases } from 'data/columns/columnAliasing';
 
 interface Props extends PanelProps<DatatableOptions> { }
