+++ src/data/cells/cellRenderer.test.ts
@@ -1,7 +1,7 @@
 /**
  * Tests for Rendering a Cell
  */
-import { ColumnStyleItemType, ColumnStyles, DateFormats } from 'types';
+import { ColumnStyleItemType, ColumnStyles, DateFormats, FormattedColumnValue } from 'types';
 import {
   applyFormat,
   FormatColumnValue,
@@ -13,7 +13,6 @@ import {
   TimeFormatter,
 } from './cellRenderer';
 import { Field, FieldConfig, FieldType, GrafanaTheme2, TimeRange, dateTime} from '@grafana/data';
-import { FormattedColumnValue } from '../types';
 // Single reference epoch reused across FormatColumnValue / TimeFormatter
 // cases so the human-readable date is declared once at the top of the file
 // instead of inferred from each assertion's `.toBe('2025-04-12 …')`.
