+++ src/data/cells/createdCellHelpers.ts
@@ -1,7 +1,6 @@
 import { getCellColors } from "./cellColors";
-import { DTColumnType, DTData, FormattedColumnValue } from "../types";
+import { ColumnStyleItemType, ColumnStyles, DTColumnType, DTData, FormattedColumnValue } from "types";
 import { ProcessClickthrough } from "./cellRenderer";
-import { ColumnStyleItemType, ColumnStyles } from "types";
 import { InterpolateFunction, TimeRange } from "@grafana/data";
