+++ src/data/columns/columnStyles.test.ts
@@ -2,9 +2,8 @@
  * Tests for Column Styling
  */
 
-import { ColumnStyleItemType, ColumnStyles, DateFormats } from 'types';
+import { ColumnStyleItemType, ColumnStyles, DateFormats, DTColumnType } from 'types';
 import { ApplyColumnStyles } from './columnStyles';
-import { DTColumnType } from '../types';
 
 describe('Column Styles', () => {
   const columns: DTColumnType[] = [
