+++ src/data/columns/columnStyles.ts
@@ -1,5 +1,4 @@
-import { ColumnStyleItemType } from "types";
-import { DTColumnType } from "../types";
+import { ColumnStyleItemType, DTColumnType } from "types";
 
 export const ApplyColumnStyles = (columns: DTColumnType[], columnStyles: ColumnStyleItemType[]) => {
   for (const item of columns) {
