--- src/data/types.ts
@@ -1,25 +1,0 @@
-import { FieldConfig } from "@grafana/data";
-import { ColumnStyleItemType } from "types";
-
-export interface DTColumnType {
-  title: string;
-  data: string;
-  type: string;
-  className: string;
-  fieldConfig?: FieldConfig<any>;
-  columnStyles: ColumnStyleItemType[];
-  widthHint?: string;
-  visible: boolean;
-};
-
-export interface FormattedColumnValue {
-  valueRaw: number | string | null,
-  valueFormatted: string,
-  valueRounded: number | null,
-  valueRoundedAndFormatted: string | null,
-};
-
-export interface DTData {
-  Columns: DTColumnType[];
-  Rows: any[];
-};
